requirements.txt
cohere
requests
python-dotenv
